Add Util into Context
