+++ test/shopify-cli/monorail_test.rb
@@ -25,13 +25,13 @@ module ShopifyCli
     end
 
     def test_log_does_not_write_when_not_system
-      Util.stubs(:system?).returns(false)
+      ShopifyCli::Context.any_instance.stubs(:system?).returns(false)
       Monorail.log.invocation('test', ['arg'])
       assert_equal 0, @events.size
     end
 
     def test_log_promopts_for_consent
-      Util.stubs(:system?).returns(true)
+      ShopifyCli::Context.any_instance.stubs(:system?).returns(true)
       Config.stubs(:get_section).returns({})
       CLI::UI::Prompt.expects(:confirm).returns(true)
       Config.expects(:set).with('analytics', 'enabled', true)
@@ -39,7 +39,7 @@ module ShopifyCli
     end
 
     def test_log_writes_to_file_when_enabled
-      Util.stubs(:system?).returns(true)
+      ShopifyCli::Context.any_instance.stubs(:system?).returns(true)
       Config.stubs(:get).returns('true')
       ShopifyCli::Log.any_instance.expects(:write)
       Monorail.log.invocation('test', ['arg'])
@@ -54,7 +54,7 @@ module ShopifyCli
 
     def test_send_events_sends_when_enabled_and_consented
       write_events
-      Util.stubs(:system?).returns(true)
+      ShopifyCli::Context.any_instance.stubs(:system?).returns(true)
       Config.stubs(:get).returns('true')
       Monorail.expects(:produce)
       Monorail.send_events
@@ -63,7 +63,7 @@ module ShopifyCli
     def test_send_events_sends_only_new_events
       write_events(3)
       File.write(@mtime, @_events.first[:timestamp])
-      Util.stubs(:system?).returns(true)
+      ShopifyCli::Context.any_instance.stubs(:system?).returns(true)
       Config.stubs(:get).returns('true')
       Monorail.expects(:produce).twice
       Monorail.send_events
